Add API get preferences folders
